# docs/openapi.yaml
openapi: 3.0.3
info:
  title: cardinal-wings API
  description: >
    REST API daemon for managing cardinal over HTTP. Panel-oriented surface
    over cardinal's Docker-compatible API, with multi-node routing (`?node=`),
    async tasks, streaming logs/exec, metrics and role-based auth.
  version: "0.2.0-dev"
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
...
paths:
...
  /download/file:
    get:
      tags: [system]
      summary: Browser file download via panel-signed JWT (?token=, scope file-download)
      parameters: [{ name: token, in: query, required: true, schema: { type: string } }]
      responses: { "200": { description: file bytes } }
  /upload/file:
    post:
      tags: [system]
      summary: Browser file upload via panel-signed JWT (?token=, scope file-upload), multipart files
      parameters: [{ name: token, in: query, required: true, schema: { type: string } }]
      responses: { "200": { description: uploaded } }
  /download/backup:
    get:
      tags: [system]
      summary: Stored snapshot download via panel-signed JWT (?token=, scope backup-download)
      parameters: [{ name: token, in: query, required: true, schema: { type: string } }]
      responses: { "200": { description: tar.gz snapshot } }
  /api/servers/{uuid}/ws:
    get:
      tags: [system]
      summary: Legacy console websocket protocol (panel JWT auth frame, then status/console output/stats)
      parameters: [{ name: uuid, in: path, required: true, schema: { type: string } }]
      responses: { "101": { description: switching protocols } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...